API: change Divergence init pattern and move move differential operators to diff_ops.py
Changes in detail:
- move finite_diff, PartialDerivative, Gradient, Divergence, Laplacian
  to tensor_ops.py
- make PartialDerivative, Gradient, Divergence, Laplacian subclass of
  PointwiseTensorFieldOperator
- implement adjoint of PartialDerivative and add a test for it
- allow Gradient and Divergence to be initialized with domain or range
  (or both); add doctests for those init patterns
- use global tuple of supported difference schemes instead of a hard-coded tuple
