+++ odl/discr/__init__.py
@@ -47,3 +47,6 @@ __all__ += discr_ops.__all__
 
 from .tensor_ops import *
 __all__ += tensor_ops.__all__
+
+from .diff_ops import *
+__all__ += diff_ops.__all__
